feat: support default values for query parameters
