+++ packages/gatsby-theme-api-docs/src/components/type/properties/rows/description.js
@@ -1,11 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { css } from '@emotion/react';
-import {
-  Markdown,
-  designSystem,
-  useISO310NumberFormatter,
-} from '@commercetools-docs/ui-kit';
+import { Markdown, designSystem } from '@commercetools-docs/ui-kit';
 import SpacingsInline from '@commercetools-uikit/spacings-inline';
 import SpacingsStack from '@commercetools-uikit/spacings-stack';
 import extractAdditionalInfo from '../../../../utils/extract-additional-info';
@@ -13,6 +9,7 @@ import capitalizeFirst from '../../../../utils/capitalize-first';
 import { useApiTypeByApiKeyAndDisplayName } from '../../../../hooks/use-api-types';
 import { DescriptionText } from '../../../description';
 import Info from '../../../info';
+import InfoValue from '../../../info-value';
 
 const customCodeStyle = css`
   font-size: ${designSystem.typography.fontSizes.small};
@@ -62,46 +59,6 @@ ConstantLikeEnumDescription.propTypes = {
   }),
 };
 
-const InfoValue = (props) => {
-  const value = props.children;
-  const valueType = typeof value;
-  const formatNumber = useISO310NumberFormatter();
-
-  switch (valueType) {
-    case 'boolean':
-      return value ? (
-        ''
-      ) : (
-        <>
-          :{' '}
-          <Markdown.InlineCodeWithoutBox css={customCodeStyle}>
-            No
-          </Markdown.InlineCodeWithoutBox>
-        </>
-      );
-    case 'number':
-      return (
-        <>
-          :{' '}
-          <Markdown.InlineCodeWithoutBox css={customCodeStyle}>
-            {formatNumber(value)}
-          </Markdown.InlineCodeWithoutBox>
-        </>
-      );
-    default:
-      return (
-        <>
-          :{' '}
-          <Markdown.InlineCodeWithoutBox css={customCodeStyle}>
-            {value}
-          </Markdown.InlineCodeWithoutBox>
-        </>
-      );
-  }
-};
-InfoValue.propTypes = {
-  children: PropTypes.any.isRequired,
-};
 const AdditionalInfo = (props) => {
   const additionalInfos = extractAdditionalInfo(props.property);
   return (
