+++ packages/gatsby-theme-api-docs/src/hooks/use-api-resources.js
@@ -14,6 +14,7 @@ export const useApiResources = () => {
         queryParameters {
           name
           required
+          default
           type
           builtinType
           description
