+++ packages/gatsby-transformer-raml/src/schema/define-raml-resource.mjs
@@ -91,6 +91,7 @@ const defineRamlResource = ({ schema, createTypes }) => {
       fields: {
         name: 'String!',
         required: 'Boolean',
+        default: 'String',
         type: 'String!',
         builtinType: 'String!',
         unionParams: '[RamlResourceQueryUnionParameter!]',
